fix(web): split Channel types into client-safe module
`channel-list.tsx` is a client component. Importing `channelSlug`
(a value) from `emotes-service.ts` pulled the whole module into the
client bundle, including the top-level `getEnvOrThrow` reads for
`EMOTES_SERVICE_API_*`. Those env vars are server-only, so the
bundle threw on hydration and the channel cards failed to render.

Move the discriminated union, `channelSlug` and `GLOBAL_CHANNEL` into
`lib/api/channels.ts` (no env reads). `emotes-service.ts` consumes
them server-side; the component imports from the pure module.
