+++ apps/web/lib/api/emotes-service.ts
@@ -1,4 +1,11 @@
 import { getEnvOrThrow } from '~/lib/helpers';
+import {
+  GLOBAL_CHANNEL,
+  type Channel,
+  type TwitchChannel,
+} from '~/lib/api/channels';
+
+export type { Channel } from '~/lib/api/channels';
 
 const apiUrl = getEnvOrThrow('EMOTES_SERVICE_API_URL');
 const apiKey = getEnvOrThrow('EMOTES_SERVICE_API_KEY');
@@ -45,26 +52,6 @@ export const getActiveEmotes = (channel: string): Promise<ActiveEmote[]> =>
 export const getRemovedEmotes = (channel: string): Promise<RemovedEmote[]> =>
   fetchEmotes(`${apiUrl}/v1/emotes/${channel}/removed`, 'getRemovedEmotes');
 
-export interface GlobalChannel {
-  type: 'global';
-  id: 'global';
-  displayName: string;
-  icon: string;
-}
-
-export interface TwitchChannel {
-  type: 'twitch';
-  id: string;
-  twitchId: string;
-  displayName: string;
-  imageUrl: string;
-}
-
-export type Channel = GlobalChannel | TwitchChannel;
-
-export const channelSlug = (channel: Channel): string =>
-  channel.type === 'global' ? 'global' : channel.twitchId;
-
 interface ChannelDto {
   id: string;
   twitchId: string;
@@ -72,18 +59,12 @@ interface ChannelDto {
   imageUrl: string;
 }
 
-const GLOBAL_CHANNEL: GlobalChannel = {
-  type: 'global',
-  id: 'global',
-  displayName: 'Global',
-  icon: '🌐',
-};
-
 export const getChannels = async (): Promise<Channel[]> => {
   const channels = await fetchEmotes<ChannelDto[]>(
     `${apiUrl}/v1/channels`,
     'getChannels'
   );
+
   const twitchChannels: TwitchChannel[] = channels.map((c) => ({
     type: 'twitch',
     id: c.id,
